+++ lib/screen/notification/provider/notification_notifier.dart
@@ -1,30 +1,35 @@
 import 'package:book_brain/screen/notification/service/notification_service.dart';
 import 'package:book_brain/service/api_service/response/notification_response.dart';
 import 'package:book_brain/utils/core/base/base_notifier.dart';
+import 'package:book_brain/utils/core/helpers/auth_helper.dart';
 
-class NotificationNotifier extends BaseNotifier{
+class NotificationNotifier extends BaseNotifier {
   NotificationService notifications = NotificationService();
   List<NotificationResponse>? listNotifications = [];
 
-
   Future<void> getData() async {
+    if (!AuthHelper.isLoggedIn) return;
     await getNotification();
   }
 
-
   Future<bool> getNotification() async {
-    return await execute(() async{
-      listNotifications = await notifications.getListNotification(page: 1, limit: 10, unreadOnly: false);
+    if (!AuthHelper.isLoggedIn) return false;
+    return await execute(() async {
+      listNotifications = await notifications.getListNotification(
+        page: 1,
+        limit: 10,
+        unreadOnly: false,
+      );
       notifyListeners();
       print("listNotifications $listNotifications");
       return true;
     });
   }
 
-
   Future<bool> deleteNotification(int notificationId) async {
-    return await execute(() async{
-       await notifications.deleteNotification(notificationId: notificationId);
+    if (!AuthHelper.isLoggedIn) return false;
+    return await execute(() async {
+      await notifications.deleteNotification(notificationId: notificationId);
       notifyListeners();
       print("listNotifications $listNotifications");
       return true;
@@ -32,7 +37,8 @@ class NotificationNotifier extends BaseNotifier{
   }
 
   Future<bool> deleteAllNotification() async {
-    return await execute(() async{
+    if (!AuthHelper.isLoggedIn) return false;
+    return await execute(() async {
       await notifications.deleteAllNotification();
       notifyListeners();
       print("listNotifications $listNotifications");
@@ -41,7 +47,8 @@ class NotificationNotifier extends BaseNotifier{
   }
 
   Future<bool> markNotification(int notificationId) async {
-    return await execute(() async{
+    if (!AuthHelper.isLoggedIn) return false;
+    return await execute(() async {
       await notifications.markReadNotification(notificationId: notificationId);
       notifyListeners();
       print("listNotifications $listNotifications");
@@ -50,7 +57,8 @@ class NotificationNotifier extends BaseNotifier{
   }
 
   Future<bool> markAllNotification() async {
-    return await execute(() async{
+    if (!AuthHelper.isLoggedIn) return false;
+    return await execute(() async {
       await notifications.markReaAllNotification();
       notifyListeners();
       print("listNotifications $listNotifications");
